+++ packages/agtree/rollup.config.ts
@@ -10,6 +10,7 @@ import externals from 'rollup-plugin-node-externals';
 import dtsPlugin from 'rollup-plugin-dts';
 import alias from '@rollup/plugin-alias';
 import json from '@rollup/plugin-json';
+import replace from '@rollup/plugin-replace';
 import path from 'node:path';
 import { readFileSync } from 'node:fs';
 
@@ -20,6 +21,7 @@ const ROOT_DIR = __dirname;
 const BASE_FILE_NAME = 'agtree';
 const BASE_NAME = 'AGTree';
 const PKG_FILE_NAME = 'package.json';
+const COMPATIBILITY_TABLE_FILENAME = 'compatibility-table-data.js';
 
 const distDir = path.join(ROOT_DIR, 'dist');
 const pkgFileLocation = path.join(ROOT_DIR, PKG_FILE_NAME);
@@ -52,13 +54,12 @@ const banner = `/*
 
 // Common plugins for all types of builds
 export const commonPlugins = [
-    json({ preferConst: true }),
     alias({
         entries: [
             // Replace dynamic compatibility table data builder with the pre-built data file
             {
                 find: './compatibility-table-data',
-                replacement: path.resolve(ROOT_DIR, 'dist/compatibility-tables.json'),
+                replacement: path.resolve(distDir, COMPATIBILITY_TABLE_FILENAME),
             },
             // Add ".js" extension to all imports of the "semver" package, eg "semver/functions/..."
             // We need this because we import functions from the "semver" package directly,
@@ -73,9 +74,10 @@ export const commonPlugins = [
             },
         ],
     }),
-    commonjs({ sourceMap: false }),
-    resolve({ preferBuiltins: false }),
     externals(),
+    json({ preferConst: true }),
+    resolve({ preferBuiltins: false }),
+    commonjs({ sourceMap: false }),
     typescript({
         tsconfig: path.join(ROOT_DIR, 'tsconfig.json'),
         compilerOptions: {
@@ -85,9 +87,40 @@ export const commonPlugins = [
         },
         include: [path.join(ROOT_DIR, './src/**/*.ts')],
         exclude: [path.join(ROOT_DIR, './node_modules'), path.join(ROOT_DIR, './test')],
-        outputToFilesystem: true,
+        outputToFilesystem: false,
+    }),
+    replace({
+        preventAssignment: true,
+        delimiters: ["'", "'"],
+        values: {
+            [path.join(distDir, COMPATIBILITY_TABLE_FILENAME)]: `'./${COMPATIBILITY_TABLE_FILENAME}'`,
+        },
     }),
 ];
+
+const compatibilityTablesBanner = `/**
+* @file Compatibility tables data for AGTree
+*
+* This file is auto-generated from YAML files in the "compatibility-tables" directory.
+* It is optimized for better runtime usage and storage efficiency.
+*
+* We use "shared" section to share the same values between different map keys
+* to reduce the storage usage.
+*/`;
+
+const compatibilityTables = {
+    input: path.join(distDir, 'compatibility-table-data.json'),
+    output: [
+        {
+            file: path.join(distDir, COMPATIBILITY_TABLE_FILENAME),
+            format: 'cjs',
+            exports: 'named',
+            sourcemap: false,
+            banner: compatibilityTablesBanner,
+        },
+    ],
+    plugins: [json()],
+};
 
 const node = {
     input: path.join(ROOT_DIR, 'src/index.ts'),
@@ -105,6 +138,9 @@ const node = {
             sourcemap: false,
             banner,
         },
+    ],
+    external: [
+        path.resolve(distDir, COMPATIBILITY_TABLE_FILENAME),
     ],
     plugins: commonPlugins,
 };
@@ -127,4 +163,4 @@ const dts = {
 };
 
 // Export build configs for Rollup
-export default [node, dts];
+export default [compatibilityTables, node, dts];