+++ packages/css-tokenizer/test/smoke/esm/index.js
@@ -1,0 +1,12 @@
+import { tokenize } from '@adguard/css-tokenizer';
+import { ok } from 'assert';
+
+let tokensCount = 0;
+
+tokenize('div { color: red; }', () => {
+    tokensCount += 1;
+});
+
+ok(tokensCount > 0);
+
+console.log('Smoke test passed');